Fixed issues with angles. Added flightpath to represent drone moves and exception handling for paths being too long.

+++ src/main/java/uk/ac/ed/inf/DroneMove.java
@@ -2,5 +2,5 @@ package uk.ac.ed.inf;
 
 
 public record DroneMove(String orderNo, double fromLongitude, double fromLatitude,
-                        int angle, double toLongitude, double toLatitude) {
+                        double angle, double toLongitude, double toLatitude) {
 }
